requirements.txt
Pillow~=9.0.1
opencv-python~=4.5.5.62
future~=0.18.2
torch~=1.10.2
utils~=1.0.1
numpy~=1.21.5
pandas~=1.1.5
scipy~=1.7.3
keras~=2.8.0
imutils~=0.5.4
sklearn~=0.0
scikit-learn~=1.0.2
tensorflow~=2.8.0
efficientnet_pytorch~=0.7.1